+++ lang/en/app.php
@@ -1,6 +1,0 @@
-<?php
-
-return [
-    'title' => 'USIM Application',
-    'welcome' => 'Welcome, :name',
-];