+++ lang/en/auth.php
@@ -1,8 +1,0 @@
-<?php
-
-return [
-    'login' => [
-        'description' => 'Enter your credentials to continue.',
-        'title' => 'Sign in to your account',
-    ],
-];