+++ lang/en/media.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'logo' => 'USIM logo',
-];