+++ lang/en/ui.screens.admin.dashboard.php
@@ -1,7 +1,0 @@
-<?php
-
-return [
-    'add_user' => 'Add user',
-    'dashboard' => 'Dashboard',
-    'search_users' => 'Search users...',
-];