+++ lang/en/ui.screens.demo.select_demo.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'all_selections_have_' => 'All selections have been reset. Start over!',
-];