+++ lang/es/app.php
@@ -1,6 +1,0 @@
-<?php
-
-return [
-    'title' => 'Aplicacion USIM',
-    'welcome' => 'Bienvenido, :name',
-];