+++ lang/es/auth.php
@@ -1,8 +1,0 @@
-<?php
-
-return [
-    'login' => [
-        'description' => 'Ingresa tus credenciales para continuar.',
-        'title' => 'Inicia sesion en tu cuenta',
-    ],
-];