+++ lang/es/media.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'logo' => 'Logo de USIM',
-];