+++ lang/es/ui.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'empty_state' => 'No se encontraron registros',
-];