+++ lang/es/ui.screens.demo.select_demo.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'all_selections_have_' => 'Las selecciones han sido reseteadas. Comience!',
-];